[ALPS03682533] platform: fix 32 bits project build break
Modify kernel config and fix coding issue

MTK-Commit-Id: 5f24348f68452c6990e53016ea236a73902b078a

CR-Id: ALPS03682533
Feature: Others

+++ drivers/misc/mediatek/ccu/src/mt6765/ccu_hw.c
@@ -1091,10 +1091,10 @@ int ccu_run(void)
 	 *     query INFO00 & INFO01 as mailbox address
 	 */
 	pMailBox[MAILBOX_SEND] =
-		(struct ccu_mailbox_t *)(dmem_base +
+		(struct ccu_mailbox_t *)(uintptr_t)(dmem_base +
 			ccu_read_reg(ccu_base, CCU_DATA_REG_MAILBOX_CCU));
 	pMailBox[MAILBOX_GET] =
-		(struct ccu_mailbox_t *)(dmem_base +
+		(struct ccu_mailbox_t *)(uintptr_t)(dmem_base +
 			ccu_read_reg(ccu_base, CCU_DATA_REG_MAILBOX_APMCU));
 
 
@@ -1275,7 +1275,7 @@ int ccu_i2c_ctrl(unsigned char i2c_write_id, int transfer_len)
 
 int ccu_read_info_reg(int regNo)
 {
-	int *offset = (int *)(ccu_base + 0x60 + regNo * 4);
+	int *offset = (int *)(uintptr_t)(ccu_base + 0x60 + regNo * 4);
 
 	LOG_DBG("ccu_read_info_reg: %x\n", (unsigned int)(*offset));
